+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changed
 
 - Simplify the timeline visualization in replay mode.
+- Disable the `onNewSchedule` for the Fray debugger plugin to improve the performance.
 
 ### Deprecated